modules: Fix signdness warning

+++ src/lib/modules.c
@@ -170,7 +170,7 @@ __imlib_ModuleFind(char **ppath, const char *name)
      {
         n = snprintf(nbuf, sizeof(nbuf), "%s/%s.so", *pp, name);
 
-        if (n < 0 || n >= sizeof(nbuf) || !__imlib_FileIsFile(nbuf))
+        if (n < 0 || n >= (int)sizeof(nbuf) || !__imlib_FileIsFile(nbuf))
            continue;
 
         return strdup(nbuf);
